[ISSUE#3179]Replace the NotifyCenter with new refactor in the config and test module.
