+++ config/src/main/java/com/alibaba/nacos/config/server/model/event/DerbyImportEvent.java
@@ -16,14 +16,14 @@
 
 package com.alibaba.nacos.config.server.model.event;
 
-import com.alibaba.nacos.core.notify.SlowEvent;
+import com.alibaba.nacos.common.notify.SlowEvent;
 
 /**
  * Data import event.
  *
  * @author <a href="mailto:liaochuntao@live.com">liaochuntao</a>
  */
-public class DerbyImportEvent implements SlowEvent {
+public class DerbyImportEvent extends SlowEvent {
     
     private static final long serialVersionUID = 3299565864352399053L;
