+++ config/src/main/java/com/alibaba/nacos/config/server/model/event/DerbyLoadEvent.java
@@ -16,14 +16,14 @@
 
 package com.alibaba.nacos.config.server.model.event;
 
-import com.alibaba.nacos.core.notify.SlowEvent;
+import com.alibaba.nacos.common.notify.SlowEvent;
 
 /**
  * DerbyLoadEvent.
  *
  * @author <a href="mailto:liaochuntao@live.com">liaochuntao</a>
  */
-public class DerbyLoadEvent implements SlowEvent {
+public class DerbyLoadEvent extends SlowEvent {
     
     public static final DerbyLoadEvent INSTANCE = new DerbyLoadEvent();
