+++ config/src/main/java/com/alibaba/nacos/config/server/model/event/RaftDbErrorRecoverEvent.java
@@ -17,7 +17,7 @@
 package com.alibaba.nacos.config.server.model.event;
 
 import com.alibaba.nacos.common.JustForTest;
-import com.alibaba.nacos.core.notify.Event;
+import com.alibaba.nacos.common.notify.Event;
 
 /**
  * RaftDBErrorRecoverEvent.
@@ -25,6 +25,6 @@ import com.alibaba.nacos.core.notify.Event;
  * @author <a href="mailto:liaochuntao@live.com">liaochuntao</a>
  */
 @JustForTest
-public class RaftDbErrorRecoverEvent implements Event {
+public class RaftDbErrorRecoverEvent extends Event {
 
 }
